Модули создания и вставки ноды объединены в модуль для служебных функций utils
