+++ js/main.js
@@ -1,4 +1,4 @@
 import introElement from './intro.js';
-import insertBlock from './page.js';
+import {insertBlock} from './utils.js';
 
 insertBlock(introElement);
