+++ js/rules.js
@@ -1,7 +1,7 @@
 /**
  * Created by Viacheslav on 20.11.2016.
  */
-import getElementFromTemplate from './getElement.js';
+import {getElementFromTemplate} from './utils.js';
 import nextScreen from './game.js';
 
 const rulesElement = getElementFromTemplate(
