+++ js/templates/view.js
@@ -1,0 +1,15 @@
+export default class AbstractView {
+
+  getMarkup() {
+
+  }
+
+  bindHandlers() {
+
+  }
+
+  clearHandlers() {
+
+  }
+
+}
